+++ core/controlador/controlador.py
@@ -1,0 +1,2 @@
+class ControladorJuego:
+    def __init__(self)